#16 opeing/1 컴포넌트 분리, SearchResultComponent Init
