+++ src/components/opening/common/HeaderComponent.tsx
@@ -1,7 +1,7 @@
 import styled from "styled-components";
 import Link from "next/link";
 import { useRecoilState } from "recoil";
-import { nextStepState, currentStepState } from "../states";
+import { currentStepState } from "../states";
 
 const HeaderWrapper = styled.div`
   display: flex;
@@ -23,32 +23,6 @@ const ArrowLeft = styled.div`
   left: 20px;
   cursor: pointer;
 `;
-// const Progress = styled.svg`
-//   position: absolute;
-//   z-index: auto;
-//   color: #ffd262;
-//   .progress-container {
-//     stroke: #f1f1f3;
-//   }
-//   .progress-content {
-//     stroke: #ffd262;
-//   }
-
-//   .progress-content > line {
-//     stroke-dasharray: 100%;
-//     stroke-dashoffset: ${(props: { dashoffset: number }) => props.dashoffset}%;
-//     animation: svg-progress 2s ease-out;
-//   }
-//   @keyframes svg-progress {
-//     0% {
-//       stroke-dashoffset: 100%;
-//     }
-//     100% {
-//       stroke-dashoffset: ${(props: { dashoffset: number }) =>
-//         props.dashoffset}%;
-//     }
-//   }
-// `;
 const Progress = styled.progress`
   -webkit-appearance: none;
   -moz-appearance: none;
@@ -72,37 +46,22 @@ export default function IndexComponent(): JSX.Element {
   return (
     <>
       <HeaderWrapper>
-        <ArrowLeft>
-          {currentStep !== 1 ? (
-            <a onClick={() => setCurrentStep(currentStep - 1)}>
-              <img src="/assets/search/arrowleft.svg" />
-            </a>
-          ) : (
-            <Link href="/main/books">
+        {currentStep !== 1 ? (
+          <ArrowLeft onClick={() => setCurrentStep(currentStep - 1)}>
+            <img src="/assets/search/arrowleft.svg" />
+          </ArrowLeft>
+        ) : (
+          <Link href="/main/books">
+            <ArrowLeft>
               <a>
                 <img src="/assets/search/arrowleft.svg" />
               </a>
-            </Link>
-          )}
-        </ArrowLeft>
+            </ArrowLeft>
+          </Link>
+        )}
         <div>스터디 주최</div>
       </HeaderWrapper>
       <Progress value={20 * currentStep} max="100" />
-      {/* <Progress dashoffset={props.dashoffset} height="375" width="375">
-        <g className="progress-container">
-          <line x1="0" y1="0" x2="100%" y2="0" strokeWidth="2" />
-        </g>
-        <g className="progress-content">
-          <line
-            x1="0"
-            y1="0"
-            x2="100%"
-            y2="0"
-            fill="transparent"
-            strokeWidth="2"
-          />
-        </g>
-      </Progress> */}
     </>
   );
 }
