+++ src/components/opening/states.tsx
@@ -1,5 +1,6 @@
 import { atom } from "recoil";
 
+// common
 const currentStepState = atom({
   key: `currentStep`,
   default: 1,
@@ -8,8 +9,14 @@ const nextStepState = atom({
   key: `nextStep`,
   default: false,
 });
+// 1
+const categoryState = atom({
+  key: `category`,
+  default: 0,
+});
+// 2
 const typeState = atom({
   key: `type`,
   default: "",
 });
-export { currentStepState, nextStepState, typeState };
+export { currentStepState, nextStepState, categoryState, typeState };
